Remove obsolete and premature libproc env var toggles.
libproc has, or rather had, a number of toggles that could be flipped via
environment variables: _LIBPROC_INCORE_ELF forced all ELF objects to be read
in-core via the 'fake ELF' mechanism, and LIBPROC_NO_QSORT forced linear
searches of all symbol tables on the grounds that only a few lookups would be
required and the symbol table was huge so the sorting would dominate runtime.

The former flag is obsolete after commit 64ba658's wholesale removal of fake ELF
files; the latter flag is a ridiculous premature optimization when viewed from
the perspective of a runtime tracer like DTrace. and not worth the hundred-plus
lines of code it takes to implement it.  (It made more sense in the original
libproc, which could perfectly well be called for binaries with vast symbol
tables only to do a single lookup and immediately exit. Indeed the p*(1) tools
in Solaris do pretty much exactly that.  But DTrace runs for a long time; symbol
table sorting will never dominate its runtime, nor be a noticeable part of it.
DTrace never sets this flag, and it is highly unlikely that any of its users
would either.)