Cater for rtld_db-iterated entities with no name.
Some of the things rtld_db iterates over are not normal shared libraries, and
have no SONAME, hence no name in ld.so's link maps.  (These include the
executable itself, and the vdso.)

Armour the map iterator in Psymtab againt this case.  (A later commit will
add a mechanism to tell which is which.)