Let internal callers of Pxlookup_by_name() suppress load address adjustment.
Pxlookup_by_name() always adjusts the address of (non-TLS) symbols it looks up
by the file_dyn_base for the file it is found in.  This adjusts for the
difference between ELF file addresses and addresses in memory -- but a
forthcoming change requires a name lookup in order to *determine* this value,
and we don't want to trigger such adjustments in that case.

So turn Pxlookup_by_name() back into a wrapper function and introduce
Pxlookup_by_name_internal(), which has a new 'fixup_load_addr' parameter, which
the wrapper sets to 1: if 0, all load-address adjustment is suppressed.