Correctly compute the file_dyn_base for relocatable objects, and ld.so.
There were two problems with file_dyn_base computation which combined to destroy
the utility of name lookups anywhere but non-PIE executables (you got answers
back, but all the addresses were wrong).

Firstly, the rl_base provided by rtld_db is not a *correction* to the values
given in the ELF files: it is a replacement for them, the difference between the
addresses in the ELF file and the addresses in memory.  No other adjustment for
load addresses or anything need be done.

Secondly, ld.so needs special handling, because it does not update rl_base for
itself, so we have to compute its file_dyn_base some other way.  If we can look
up the _r_debug symbol in the ELF file and the value of the DT_DEBUG dynamic
tag (which *is* updated by ld.so), we can subtract the one from the other, and
the result is the file_dyn_base.  (If we can't look up _r_debug, it doesn't
matter: all our manipulations of ld.so depend on _r_debug first of all, so
without that symbol we can get nothing useful out of it.)

Of course, we have to look up _r_debug without applying file_dyn_base to it: we
can use the Pxlookup_by_name_internal() function introduced in the previous
commit.