Fix race conditions in breakpoint continuation.
When the breakpoint machinery was introduced, Puntrace() was adjusted to do all
its resumes (other than a final detach) via Pbkpt_continue() rather than via a
direct PTRACE_CONT.  This was intended to allow breakpoints to be automatically
resumed when necessary -- but Pbkpt_continue() was not up to the job.

There are multiple race conditions in Pbkpt_continue(), most notably that it
attempts to determine whether a singlestep has been done by remembering if it
asked for one.  But that will not do: the process could be hit by a signal and
halted again either before or after the instant of singlestepping, and whether
it has executed the instruction is a matter of timing.  This might seem like a
very narrow race window, but signals are not always delivered at once.  You can
easily trigger it by hitting a process currently stalled on a breakpoint with
SIGSTOP, then resuming it.

The solution is to ditch the dead-reckoning method and the 'singlestepped'
variable it implied, and look at the IP address at Pbkpt_continue() time
instead (with any post-SIGTRAP platform compensations applied, as usual).  If
it's the same as the breakpoint address, we know that the singlestep hasn't
happened yet.  If what we get back is an -ESRCH, we know that someone called
Pbkpt_continue() when we were already continued, so it turn it into a quick
non-blocking Pwait() instead, in case a breakpoint trap is waiting.

If the singlestep is done, we might still have SIGTRAPs waiting from it: the
user could have called called Pbkpt_continue() rather than choosing to Pwait(),
without knowing that the singlestep handler would have been called if they'd
Pwait()ed (due to the pending SIGTRAP).  But there might *not* be a trap (they
are not reliably delivered on many platforms if you've been hitting the process
with other signals at the same instant), so we clearly can't just leave the
execution of the breakpoint handler up to Pwait(): instead, we set a new
bkpt_consume state variable, which acts as an instruction to Pwait() to throw
away any pending breakpoint SIGTRAPs rather than invoke their handlers, then
unconditionally invoke the post-singlestep handler ourselves.  (While we're
at it, we rename that handler to bkpt_handle_post_singlestep(), to better
reflect its purpose.)