Revamp of DTRACE_DEBUG machinery: new DTRACE_DEBUG=signal.
The existing machinery used when DTRACE_DEBUG is set is very simple, amounting
to an fprintf() to stderr with a hardwired prefix whenever a dt_dprintf() call
is encountered (inside libproc, _dprintf()).

This is not ideal: each output to stderr takes out locks, adds overhead, and can
block, and when all these are combined, turning debugging on can easily change
timings enough that problems go away.

This commit adds machinery which responds to DTRACE_DEBUG=signal by allocating a
100Mb ring buffer and directing all debugging output to that buffer via glibc's
memstream machinery.  This still incurs lock overhead (because libproc and
dtrace could be emitting to the same buffer simultaneously in different
threads), but the file-write and blocking overhead is gone.  In long debugging
runs, this can reduce the overhead of debugging by 10% or more.

The debugging buffer is flushed at dtrace_close() and Prelease() time, and also
when the libdtrace consumer is hit by a specified signal (by default SIGUSR1).
This signal can be changed via the new function dtrace_debug_set_dump_sig(),
which is added to the libdtrace API.  Setting the signal to zero disables this
machinery and forces all debugging to be done the old way, no matter the value
of DTRACE_DEBUG.

The debugging buffer can be resized via the DTRACE_DEBUG_BUF_SIZE variable,
which gives a size in megabytes.

Since this machinery is quite a lot larger than the old, we move all
debugging-related definitions into dt_debug.c and dt_debug.h (they were
previously scattered between dt_impl.h, dt_open.c and dt_subr.c).  We also rip
out the independently-implemented debugging machinery in libproc and reimplement
it in terms of a call to the same routine in dt_debug.c now used for
dt_dprintf().  This routine accepts a customizable subsystem label so that
libproc output can, as before, be differentiated from libdtrace output.