Fix false positives from unittest/libproc/tst.consistency.sh.
This test was FPing because it was not taking into account the degree to which
iteration over link maps with lmid > 0 slows the child process down.  It checks
that a "reasonable number" of dlopen()/dlclose() calls were seen by the rd_event
machinery.  A normal process doing nothing else can easily do hundreds to
thousands of these in the twenty-second window we give it, even when slowed by
periodic halts when it tries to start a dlopen()/dlclose() while link map
iteration is under way, but when nonzero lmids are in use the process is halted
for much of the duration of each iteration, so can complete far fewer dlopen()s
if it's being stopped by a constant loop of iterations.

So split the consistency test into two (differentiated by an environment
variable set by the test .sh file), one which uses one lmid, and a high count of
expected dlopen()s, and one which uses multiple lmids and a much lower count of
expected dlopen()s.  (The much lower expected count is still more than high
enough to spot a failure to spot dlopen()s at all, while being much lower than
the normal count I see here.)

We also adjust the hack inserted into libproc-consistency.c to prevent flooding
the test log with huge amounts of iteration output. For multi-lmid testing, it
sufficed to suppress printed output if the lmids fell to a value below any we
had seen before: for single-lmids testing, this will not do, and we instead
suppress output using the highest value seen for the 'num' parameter to the
iteration function, which will naturally fall back to zero after the first
iteration.  (Thus we test that explicitly as well, though it is used by
machinery in Psymtab and so is tested by other dtrace tests in any case.)