Use the symbol search scope to find symbols.
Now that we have access to the dynamic linker's symbol search scopes,
it remains to use them in symbol lookups.

We want to use them in a fairly limited way.  A lookup for bar in module
foo should not follow search scopes, lest everything in libc appear in
almost every module.  And if we want to look up a name in a nonstandard
lmid, we don't want the lookup to fail jwut because the main program
cannot see it.

So we implement a more complex three-pass procedure for global (PR_OBJ_EVERY)
lookups.  First, we look in the main executable: then, we follow its search
path: then, we do what DTrace used to do every time and look across everything
in order regardless of search paths.  This avoids the horrible problems of
looking across everything all the time (basically random results whenever you
had more than one symbol depending on which one was earlier in the address space
and so got filled in by Pupdate_maps() first), while still making it possible to
trace names that the main program cannot see because they are in dlmopen()ed
libraries.

(This procedure is not quite ideal, because if dlmopen()ed libraries are
themselves making use of symbol interposition, we will ignore that.  We should
perhaps add another pass before the 'everything in order' pass, which looks in
every lmid-leading 'default' search scope in order. But given the limited use of
dlmopen() on Linux this seems not worth spending time on yet.)

There are three parts to this problem: storing the scopes, transforming them
into a useful form, and traversing them.

Storing the scopes is done by map_iter() et al.  The scopes arrive in the
rd_loadobj_t parameter to map_iter(), but they are not useful as they are
because they will be aggressively overwritten as soon as we return: they are not
protected by the copy of the rd_loadobj_t which we take in file_info_t.file_lo
because they are separately allocated.  So we malloc() space for them in our
file_lo.rl_scopes field and copy them in.  As for when to free them -- they
share a lifetime with file_lo, not with file_info_t: the search paths can change
whenever a shared library is loaded or unloaded, rather than persisting as long
as a given library is loaded the way file_info_t does. So we free the scopes
array on file_info deletion (of course), but also on subsequent map_iter()s
and on mapping_purge() -- whenever the mappings are invalidated, so are the
search scopes.

But these scopes are still in the form of raw unadorned pointers into the
victim's address space.  This is not very useful.  However, we would rather have
them in the form of pointers to file_info_t's: the file_info_t's that correspond
to the link maps that we have victim pointers to.  This is not a terribly cheap
operation, and we're only going to be searching one scope for now (and a few in
the future, if we implement all-lmid searches): so we do it lazily, maintaining
a file_info_t.file_symsearch pointer which is NULL until Pupdate_symsearch() is
called to update it for a given file_info.

Updating it is a little fun because we don't want a change to the glibc data
structures we are using to break symbol searching: handling back an empty scope
is preferable, whereupon we fall back to the old (wrong, but
better-than-nothing) 'search in basically random order' that we had before this
commit. (The pointers themselves are safe: we know the caller will not have
changed them underneath us, because if they change a dlopen()/dlclose() has
happened and before that happens they will hit a breakpoint and we will get a
chance to invalidate our mappings.)

So Pupdate_symsearch() iterates over the scopes array, fetching each scope and
transforming it into a rd_loadobj_t with rd_get_scope(), then looking up the
address of each library's dynamic section in the mappings array just as
map_iter() does to figure out which file_info corresponds to each link map.
Any entries which cannot be figured out are simply omitted, wasting a tiny
bit of memory on the now-too-large file_symsearch array but having no other
negative consequences.

Now we have a structure containing pointers to file_info_t's which should be
searched in order.  We implement the three-phase search process by ripping out
the crude counter-based search Pxlookup_by_name_internal() was doing and
replacing it with a call to an iteration function which returns file_info_t's in
order until none are left, followed by NULL: we let it handle the one-module
case as well by adding a just_one parameter which turns the iterator into a
degenerate case that just returns the file_info you pass it and then NULL.  This
avoids turning Pxlookup_by_name_internal() into an incomprehensible mess and
makes it possible to reuse the same search semantics elsewhere later, should we
need to.

We also add a test case.  We have a 'victim' process (libproc-lookup-victim)
which links against a shared library via DT_NEEDED and also dlmopen()s it on
demand in a new link map, SIGSTOPs itself, then when resumed calls one of a
variety of functions in that library or in the main program (possibly via ELF
symbol interposition).  All those functions print error messages, because the
driver program, libproc-lookup-by-name, uses Pxlookup_by_name() to look up the
same name the victim process calls and drop a breakpoint there before SIGCONTing
it.  If the breakpoint is hit, the victim is unceremoniously killed without a
chance to recover.  So we know that Pxlookup_by_name() has resolved the wrong
address if the breakpoint is not hit or the victim manages to carry out its
function call without being killed first. The driver program is itself
repeatedly called by tests/unittest/libproc/tst.name-lookups.sh, to test main
program lookups, lookups in libraries, interposed lookups and alternate-lmid
lookups in turn.