Add a ustack() test.
This test tries to ustack() readwholedir on its first read() call.  We have to
add LDFLAGS to readwholedir to ensure it is built with a dynamic symbol table,
so the symbol lookups have a chance of success.  We also have to pause
readwholedir a bit before quitting, or there is a race window where readwholedir
can exit before dtrace userspace gets its buffers back, which also makes symbol
lookups impossible.