Post-handler breakpoint removal.
Currently, if a breakpoint handler calls Punbkpt(), the breakpoint is removed
immediately.  This was OK before the arrival of notifiers, but is seriously
problematic afterwards, because notifiers and breakpoints are supposed to be
exactly the same modulo only the ability of breakpoints to state that control is
supposed to stop rather than resuming on breakpoint return.  But we can't allow
a notifier to actually remove the breakpoint, because that would remove the list
of notifiers, including those yet to be called, and the handler and all their
associated data pointers!  So, instead, add a (per-handler) is_handler flag that
is true whenever a handler is being invoked, have Punbkpt() note in another new
per-handler flag that a removal was requested, and do it after the handlers are
finished, rather than right away.  (You can still remove *other* breakpoints
immediately from a breakpoint handler: it's only the breakpoint you're handling
that cannot be removed.)

There is one small behaviour change here: before now, it was possible for a
handler to call Punbkpt() and then return e.g. PS_STOP to nonetheless stop
execution.  You can't do that any more -- it was always rather buggy and
confused the breakpoint machinery.  You can get the same effect by returning
PS_STOP, and having code outside the breakpoint handler remove the breakpoint.
(No code in DTrace was doing this.)