Close the DTrace handle on fatal errors.
If DTrace exits with breakpoints outstanding on a child process, the child
process will die with a SIGTRAP when it hits them.  Prelease() cleans them up,
which is (eventually) called by dtrace_close(), so all exit paths from DTrace
should call this.  At least one was not: fatal(), in cmd/dtrace.c.  This appears
to be because fatal() is called at some points before DTrace is initialized, so
the DTrace handle can be NULL; the right thing to do is to make a dtrace_close()
of NULL harmless, and then call it unconditionally from fatal().