Implement the evaltime={preinit,postinit,main} option.
We do this by dropping a breakpoint on the appropriate region and getting it to
synchronize with the condition variable signalled by dt_proc_continue().

Doing this is rendered a little arcane by problems with early-stage symbol
lookup.  In particular, when dt_proc_attach() is called, the victim is still
halted just after exec(): the dynamic loader hasn't initialized and nothing has
been loaded into memory.  While we only need to look up main(), and normally
main() is in the main program, this is not actually mandatory: it can be in a
shared library.  So we need to wait until ld.so has initialized (at least, we do
if it is dynamically linked).  So we have these cases:

all: process halted at exec() by Pcreate(), resumed at Puntrace() in
     dt_proc_control(), ld.so still uninitialized

evaltime=exec: ld.so initialized in the main Pwait() loop at some unknown point
               in time

evaltime=preinit: process interrupted by breakpoint on ELF entry point (after
                  ld.so initialized, for dynamically linked programs: before,
                  for statically linked programs).  At this point, it can
                  initialize rtld_db and rendezvous with dt_proc_continue().

evaltime={postinit,main}: process interrupted by breakpoint on ELF entry point:
(same)            at this point, it can initialize rtld_db and drop a breakpoint
                  on main(), if main() exists; that breakpoint can then
                  rendezvous with dt_proc_continue().  If  there is no main(),
                  we can fall back to evaltime=preinit.

So control flow looks like this:

dt_proc_control()
  -> dt_proc_attach(.., TRUE)
    -> dt_proc_rdagent()
       # attaches to rtld_db, initializes ld.so monitoring, if possible
    -> in non-create mode, drops a breakpoint on the ELF entry point, either
       dt_break_interesting() (which does the rendezvous) or
       dt_break_drop_main() (which does the drop-on-main or failing that,
       rendezvous).
    -> if that fails, fall back to =exec mode, which imposes the fewest
       requirements.
  -> rendezvous via DT_PROC_STOP_{GRAB/CREATE}, which continues the main thread
     iff this is a grab or an =exec create.  Puntrace(), which will detach
     if and only if this is an exec-mode or grab operation with no successful
     rtld_db initialization.
  -> start Pwait()ing
    -> dt_break_drop_main() fired, when execmode=main (or =postinit).
      -> remove the breakpoint.
      -> dt_proc_attach(..., FALSE)
        -> dt_proc_rdagent() should be able to initialize now.
        -> drop dt_break_interesting() on main().
      -> if that fails, fall back to preinit mode and rendezvous,
         resuming dt_proc_continue().
    -> dt_break_interesting() fired, when execmode=main (on main()) or
       preinit (on the ELF entry point)
    -> remove the breakpoint.
    -> rendezvous, resuming dt_proc_continue().
  -> now, on every loop, do a dt_proc_rdagent().  This has no effect except
     in =exec mode, but in that mode nothing else calls dt_proc_rdagent()
     after resumption time, so it is probably not initialized.  It is
     basically free after the first call, so there is no need to check for
     failure or anything: just keep trying and eventually it will work, and
     after that we never need to do anything again.  (If there is an
     incompatibility with the bits of glibc private state we mess around with,
     rtld_db will incur a single brief timeout as part of this, but the timeout
     will not repeat.)

Testing this is a matter of triggering a binary to do something at constructor
and main() time, and dtracing to watch that something hhappen (printing
something is good enough: we can watch for the write() and watch the output get
printed).  To distinguish between =exec and =preinit, we must spy on a syscall
that the dynamic linker makes that is otherwise unlikely.  On x86-64,
arch_prctl() suffices.  We also need two binaries, one which is always stripped
and one which is never stripped and has a symbol table, to verify that the
fallback from =main to =preinit works when a symbol table is absent.