Add wrappers and hooks for multithreading support to libproc.
Currently, libproc is single-threaded and not threadsafe -- and it makes heavy
use of ptrace(), which on Linux is not even thread-aware, working only from the
single thread ptrace() was originally called in.  We would like to be able to
make it possible to transparently use libproc functions in other threads.
Rather than tiresomely proxying every single libproc operation, we'd like to
introduce a minimal number of proxies consistent with reasonable performance.

Given the Linux ptrace() restrictions, it is easiest to do this by proxying all
libproc operations to the appropriate thread, if possible transparently.  We
would also like to provide some way for the dpr_lock in dt_proc to be
automatically maintained, so that calling into a libproc function automatically
takes the lock.  And we want to do this without too much code change.

To do this, we introduce one global hook and two per-ps_prochandle wrappers.

The hook is the *Ptrace lock hook*.  This is called on outermost entry to
Ptrace() and on outermost exit from Puntrace() with appropriate arguments, and
at error points too, such that when inside any Ptrace() / Puntrace() pair the
last call to the ptrace lock hook was done with the ptracing argument equal to
1.

This necessitates an interface change.  Because Pcreate() and Pgrab() return
with the process halted by ptrace(), they must call this hook -- but almost all
ptrace lock hooks will want to pass in some extra parameter, probably one which
depends on the process.  (In our case it is the associated dt_proc structure.)
So we must adjust Pcreate() and Pgrab() to pass in that hook argument, before
the perr out argument.  While we're about it we can delete the unused len
argument from Pcreate()

The wrappers are wrappers around ptrace() and Pwait() -- the latter to get
around the kernel restriction that only the thread that ptrace()d a process can
see state changes on it via waitpid()/waitid(), and also because many state
changes require us to ptrace() anyway.  The wrappers are set to default values
at process creation / grab time, and functions in the new file wrap.c allow them
to be set on a process-by-process basis.  The wrappers should proxy the request
to the thread that did the original creation or grab, then block waiting for the
result and pass it back when it arrives.

The underlying functions that should be called are somewhat inconsistently
named, because we can minimize code change by renaming Pwait() and turning
Pwait() into a wrapper, but we cannot rename ptrace()!  So they are as
follows:

- Wrapper: Pwait(); underlying function: Pwait_internal().  This is, of course,
  a public function, so anyone may call it.

- Wrapper: wrapped_ptrace(); underlying function: ptrace().  wrapped_ptrace() is
  not a public function, though libproc is prohibited from calling ptrace()
  except via the wrapper.  So this wrapper only has an effect on ptrace() calls
  made from within libproc itself (which is where all ptrace() calls in DTrace
  are presently located).

Most of the other changes in this commit are a mass-renaming of ptrace() calls
to wrapped_ptrace().  Ptrace() itself has seen further changes, to fix problems
encountered on failed grabs, to call the ptrace lock hook, and to use Pwait()
rather than directly calling waitpid(), so that Ptrace() calls will use the
proxy for this, if present.