Do not repeatedly re-drop the r_brk dynamic debugging breakpoint.
Commit 16ac9e044f26083ed9ef3de3aecae3dd4062a900 arranged to repeatedly call
dt_proc_rdagent() on every loop through the DTrace process control thread,
because "it is basically free after the first call".  Or it should have been.
Unfortunately we forgot to set the rd_monitoring flag in this case, so on every
call we were deleting the rd_brk breakpoint and re-dropping it again *just in
case*.  Fix by setting the flag.