Two-phase dt_proc_continue().
There is a race condition in DTrace's process handling, notably that
dt_proc_continue() resumes the control thread if it's stopped at a rendezvous
event (waiting for the main thread to catch up once it's hit evaltime, or
something like that), but if the control thread *isn't* stopped at a rendezvous
event because it hasn't got to it yet, dt_proc_continue() does nothing and
shoots on, leaving the control thread stuck forever when it eventually gets
to the rendezvous.

This can lead to unbalanced Ptrace() locks and subsequenct disaster, because the
rendezvous event is right before the Puntrace() that sets the process into
detached state (it can't be afterwards because the whole point of the rendezvous
event is for the victim process to be stopped until DTrace is ready for it).
So, particularly when doing ustack()s and the like, DTrace has a tendency to
race ahead as soon sa the rendezvous is over and dive into libproc with its
new process handle, doing a Ptrace() before the control thread has got to its
Puntrace(), and then we have unbalanced Ptrace()s and disaster.

The solution is to soup up dt_proc_continue() so that it does nothing if called
after a successful resume of a stopped process, but if called before that, it
waits for the control thread to broadcast a resumption message to it.  This is
done by a new dt_proc_resume(), the counterpart of dt_proc_stop(), only where
dt_proc_stop() indicates that the control thread should stop until the main
thread is ready for it, dt_proc_resume() means that the main thread, which has
been paused since dt_proc_stop(6 returned, should resume.  We use two new
dpr_stopped states for this, DT_PROC_STOP_RESUMING (indicating that we are after
dt_proc_stop() and before dt_proc_resume()) and DT_PROC_STOP_RESUMED (indicating
that we have done a dt_proc_resume()).

There is another related race, too.  In the control thread cancellation handler,
we were looking at our fake recursive mutex owner and lock count variables to
determine whether we already had the lock out -- but when cancelled while in a
condition variable wait, the lock is automatically retaken!  So we have to add a
new flag to note when we are waiting on a condvar, and avoid retaking the lock
even though all other indications are that we don't have it.

While we're at it, fix a few places that are always called under the lock thanks
to being called from breakpoint handlers or rd activity monitors (all of which
take out the lock via the ptrace lock hook).