Clean up process destruction.
This whole area is a total mess.  My prior attempts to clean up and free dead
processes as fast as possible (in the tail of the control thread) were totally
hopeless, because the other thread could well still be holding on this process
and manipulating it, and it is interned into a hash which the main thread could
ask for it back out of at any time.

Worse yet, process death is not the only time the control thread goes away.
It's also cancelled on DTrace termination and dtrace_close() (which is process
termination in dtrace(1), but not necessarily in other consumers), and in that
situation we must be sure to remove all breakpoints and other assorted junk from
the process, and unptrace it, at thread termination time.  But right now,
Prelease() does both jobs!  (Hilariously, we were enqueuing the dpr on the
notification queue used by dtrace_sleep() right before freeing the dpr!  That's
really not going to work...)

So split the job, making Pfree() public again (as it once was before I wrongly
make it private), and making Prelease() merely release all process-related
state, and no memory, leaving a useless process in state PS_DEAD (with a flag
indicating that it's already been released so a second call to Prelease() is
harmless.).  Useless though that process is, it is not a garbage pointer and
things in the main DTrace thread can continue to use it.  (As a slight
optimization, dt_proc_release() has learnt to destroy dead processes
when called, not just retire them.  This means that most processes used by
ustack() et al are released and un-ptrace()d immediately.)

We also cannot free the various mutexes at thread cancellation: even though the
thread is gone, the main thread might still try to talk down them, and should
get an error, not a crash.

If we're not freeing everything in the control thread, where are we freeing it?
In dt_proc_destroy(), wihch we revamp to use the dpr lock functions at the same
time.  However, we should not free the dt_proc itself in there, because the
dt_proc has list pointers on it for the dph_lrulist, which our caller is busy
walking to free everything.  It's the caller that has to free it (but not
its substructure, which dt_proc_destroy can free).

(Most of this is a reversion to something more like the way Solaris DTrace did
things in this area -- but now I understand why it was this way, and have this
commit log to explain why.)