Suppress breakpoint traps during breakpoint flushing.
If, while we're shutting down and flushing all breakpoints, the victim process
hits a breakpoint trap, that trap can do things like add more breakpoints.  At
best, these will never get removed, leading the victim to SIGTRAP: more often
they will perturb the hash and lead to double-free()s or worse.

The solution is to use our existing bkpt_consume mechanism (introduced in
5a6eeb54f to handle a subtle bug with halting signals and singlestepping) to eat
and ignore all SIGTRAPs, suppressing breakpoint handlers while a flush is
underway. (Afterwards, if we are doing a Prelease() and the process is going
away, we need to do a single Pwait() to get the process going again just in case
the last breakpoint caused a SIGTRAP just before we could remove it.)