Prevent deadlocks inside breakpoint handlers.
Inside a breakpoint handler, the process is stopped, and its resumption cannot
be permitted except via a couple of controlled paths (bkpt_continue() or return
from the breakpoint handler), in order that the breakpointed instructions can be
modified.  Unfortunately Pwait() and rtld_db did not realize this, and a lot of
things indirectly call Pwait(), so it was easy to trigger a deadlock by doing the
wrong thing inside a breakpoint handler and waiting forever for the process to
resume execution when it never would because it was stopped at a breakpoint.

So, to fix this, do three things:

 - turn the bkpt_halted flag on around the execution of the breakpoint handler,
   rather than just if it returns something other than PS_RUN
 - adjust Pwait_internal() to return instantly if the bkpt_halted flag is on
 - have the rd_ldso_consistent*begin() functions return immediately if
   bkpt_halted, returning successfully if in a consistent state, and -EDEADLK
   otherwise, because if we are stopped at a breakpoint handler inside the
   dynamic linker and it is in an inconsistent state, there's no way to wait
   for it to go consistent again without resuming the process, which is
   forbidden.

In addition, fix a minor recursion bug: an outermost Puntrace() call can call
Pwait() at the end of its execution, which can (if a breakpoint is hit) recurse
back into Ptrace() again.  So, before calling Pwait(), do everything else
necessary, including flipping off the flag that says we are now halted by
Ptrace(), since by that point we no longer are.