Fix rd_event callbacks of exec()ing processes.
If a process exec()ed and had the rd_event callback enabled, the callback should
be called when the rd_start_trap is hit, since that is the first time the
complete set of DT_NEEDED libraries is visible: the dynamic linker has, in
effect, dlopen()ed them. We were not triggering the callback in this particular
case.