Support for USDT tests in testsuite, REG_ERR, and userspace sys/sdt.h.
Added support to the build system for the DTrace userspace for USDT, i.e. in
triggers.  One can now specify a provider D definition (.d file) in a
<primary>_PROV variable, and the build system will generate a matching .h file,
generate a DOF ELF object file, process all relevant object files, and link in
the DOF with the rest of the object files.

Added syscall, sdt, and usdt tests.

Added the REG_ERR register constant for the x86 architecture.

Cleaned up the drti code to not emit messages unless something goes wrong (or
if debug mode is enabled)

Added a sys/sdt.h to be installed on the system to make userspace macros
avalable for specifying USDT probes.