Handle 32-vs-64 bit correctly for dtrace -G, build 32-bit drti.o, cleanup, better tmp dir names
The dtrace command will now correctly (when possible) determine whether the
user is trying to invoke dtrace -G on 32-bit or 64-bit objects. and generate
the DOF ELF object accordingly, and use the appropriate DRTI object.

Added cleanup for provider .h files for triggers.

Changes the tst.provregex?.sh tests to use a more clear directory name for its
temporary directory (useful when a tests fails and we need to look at the
temporary files in there).