Fix symbol lookup where the first symbol found is undefined.
In this situation, Pxlookup_by_name() is meant to return the defined symbol, but
instead it is returning the first symbol found, even if that was an undefined one.

We should overwrite the output symbol not only if no overwrite has ever happened
before but also if the last time we overwrite it we overwrite it with an
undefined symbol.